feat: Add support for including Pod files in iOS and macOS builds
This commit adds the necessary changes to include Pod files in the iOS and macOS builds. It modifies the Debug.xcconfig, Release.xcconfig, Flutter-Debug.xcconfig, and Flutter-Release.xcconfig files to include the Pod files for the Runner target. It also adds the devtools_options.yaml file, which stores settings for Dart & Flutter DevTools. Additionally, it updates the pubspec.yaml file to include the flutter_svg and vector_graphics dependencies.